Support signing of PowerShell scripts (Fix #64)
