+++ README.md
@@ -13,8 +13,9 @@ or the Mono development tools on Unix systems.
 
 Jsign comes as an easy to use task/plugin for the main build systems (Maven,
 Gradle, Ant). It's especially suitable for signing executable wrappers and
-installers generated by tools like NSIS, exe4j, launch4j or JSmooth. Jsign
-can also be used programmatically or standalone as a command line tool.
+installers generated by tools like Inno Setup, NSIS, install4j, exe4j, launch4j or
+JSmooth, and PowerShell scripts.
+Jsign can also be used programmatically or standalone as a command line tool.
 
 Jsign is free to use and licensed under the Apache License version 2.0.
