+++ docs/index.html
@@ -40,16 +40,19 @@ Windows or the Mono development tools on Unix systems.</p>
 
 <p>Jsign comes as an easy to use task/plugin for the main build systems (Maven,
 Gradle, Ant). It's especially suitable for signing executable wrappers and installers
-generated by tools like <a href="http://nsis.sourceforge.net">NSIS</a>,
-<a href="http://www.ej-technologies.com/products/exe4j/overview.html">exe4j</a>,
+generated by tools like <a href="http://www.jrsoftware.org/isinfo.php">Inno Setup</a>,
+<a href="http://nsis.sourceforge.net">NSIS</a>,
+<a href="https://www.ej-technologies.com/products/install4j/overview.html">install4j</a>,
+<a href="https://www.ej-technologies.com/download/exe4j/files">exe4j</a>,
 <a href="http://launch4j.sourceforge.net">launch4j</a>
-or <a href="http://jsmooth.sourceforge.net">JSmooth</a>. Jsign can also be used
+or <a href="http://jsmooth.sourceforge.net">JSmooth</a>, and PowerShell scripts.
+Jsign can also be used
 programmatically or standalone as a command line tool.</p>
 
 <p>Jsign is free to use and licensed under the Apache License version 2.0.</p>
 
 
-<h3>Ant Task</h3>
+<h3>Ant Tasks</h3>
 
 <p>Here is an example showing how the signing works with Ant, using a Java keystore:</p>
 
@@ -75,6 +78,23 @@ programmatically or standalone as a command line tool.</p>
           keyfile="key.pvk"
           keypass="password"
           tsaurl="http://timestamp.verisign.com/scripts/timstamp.dll"/>
+</pre>
+
+<br>
+
+<p>Here is an example showing how the PowerShell signing works with Ant, using a Java keystore:</p>
+
+<pre class="prettyprint lang-xml">
+ &lt;taskdef name="signps" classname="net.jsign.PSSignerTask" classpath="jsign-2.1.jar"/>
+
+ &lt;signps file="script.ps1"
+         name="My Application"
+         url="http://www.example.com"
+         keystore="keystore.jks"
+         alias="test"
+         storepass="password"
+         tsaurl="http://timestamp.comodoca.com/authenticode"
+         scriptEncoding="ISO-8859-1"/>
 </pre>
 
 <br><br>
@@ -173,8 +193,13 @@ programmatically or standalone as a command line tool.</p>
     </tr>
     <tr>
       <td valign="top">replace</td>
-      <td valign="top">Tells if previous signatures should be replaced.</td>
+      <td valign="top">Tells if previous signatures should be replaced. (only for PESignerTask)</td>
       <td valign="top" align="center">No; defaults to "false"</td>
+    </tr>
+    <tr>
+      <td valign="top">scriptEncoding</td>
+      <td valign="top">The encoding of the script to be signed. (only for PSSignerTask)</td>
+      <td valign="top" align="center">No; defaults to "UTF-8"</td>
     </tr>
   </table>
 
@@ -184,7 +209,8 @@ programmatically or standalone as a command line tool.</p>
 <h3>Maven plugin</h3>
 
 <p>Here is an example showing how the signing works with Maven. The parameters
-are the same as those described above for the Ant task.</p>
+are the same as those described above for the Ant task. For PowerShell scripts
+use the "signps" goal instead.</p>
 
 <pre class="prettyprint lang-xml">
     &lt;build>
@@ -228,17 +254,28 @@ are the same as those described above for the Ant task.</p>
             classpath 'net.jsign:jsign-gradle-plugin:2.1'
         }
     }
-    
+
     apply plugin: 'net.jsign'
-    
-    task sign << {
-        signexe(file      : 'application.exe',
-                name      : 'My Application',
-                url       : 'http://www.example.com',
-                keystore  : 'keystore.p12',
-                alias     : 'test',
-                storepass : 'secret',
-                tsaurl    : 'http://timestamp.comodoca.com/authenticode')
+
+    task sign {
+        doLast {
+            signexe(file      : 'application.exe',
+                    name      : 'My Application',
+                    url       : 'http://www.example.com',
+                    keystore  : 'keystore.p12',
+                    alias     : 'test',
+                    storepass : 'secret',
+                    tsaurl    : 'http://timestamp.comodoca.com/authenticode')
+
+            signps(file           : 'script.ps1',
+                   name           : 'My Application',
+                   url            : 'http://www.example.com',
+                   keystore       : 'keystore.p12',
+                   alias          : 'test',
+                   storepass      : 'secret',
+                   tsaurl         : 'http://timestamp.comodoca.com/authenticode',
+                   scriptEncoding : 'ISO-8859-1')
+        }
     }
 </pre>
 
@@ -254,15 +291,15 @@ are provided to install it easily on most Linux distributions. On these systems 
 <pre> pesign [OPTIONS] FILE</pre>
 
 <p>On other systems the command line is invoked by running the jar with:</p>
-    
+
 <pre> java -jar jsign-2.1.jar [OPTIONS] FILE</pre>
-    
+
 <p>The parameters expected are the same as those used by the Ant task:</p>
 
 <pre>
   usage: pesign [OPTIONS] FILE
   Sign and timestamp a Windows executable file.
-  
+
   -s,--keystore &lt;FILE>       The keystore file, or the SunPKCS11 configuration file
      --storepass &lt;PASSWORD>  The password to open the keystore
      --storetype &lt;TYPE>      The type of the keystore:
@@ -290,6 +327,8 @@ are provided to install it easily on most Linux distributions. On these systems 
   -h,--help                  Print the help
 </pre>
 
+<p>For signing PowerShell scripts, instead use <pre>java -cp jsign-2.1.jar net.jsign.PSSignerCLI [OPTIONS] FILE</pre></p>
+
 <p>Example using a Java keystore:</p>
 
 <pre>
@@ -301,6 +340,14 @@ are provided to install it easily on most Linux distributions. On these systems 
 
 <pre>
  pesign --certfile certificate.spc --keyfile key.pvk --keypass password application.exe
+</pre>
+
+<p>Example signing a PowerShell script:</p>
+
+<pre>
+ pssign --keystore keystore.jks --alias test --storepass password \
+        --tsaurl http://timestamp.comodoca.com/authenticode \
+        --scriptEncoding ISO-8859-1 script.ps1
 </pre>
 
 
@@ -319,7 +366,7 @@ be embedded in another application.</p>
     &lt;/dependency>
 </pre>
 
-<p>and then use the <code>PESigner</code> class like this:</p>
+<p>and then use the <code>PESigner</code> or <code>PSSigner</code> class like this:</p>
 
 <pre class="prettyprint lang-java">
  KeyStore keystore = KeyStoreUtils.load(new File("keystore.p12"), "PKCS12", "password", null);
@@ -350,7 +397,7 @@ be embedded in another application.</p>
 <p>Jsign leverages the APIs developed by the <a href="http://www.bouncycastle.org">Bouncy Castle</a> project.<br>
 PVK parsing is based on the <a href="http://www.drh-consultancy.demon.co.uk/pvk.html">pvktool</a> by Stephen N Henson.</p>
 
-<p>Jsign includes contributions from Emmanuel Bourg, Florent Daigniere, Michael Szediwy, Michael Peterson, Markus Kilås and Erwin Tratar.</p>
+<p>Jsign includes contributions from Emmanuel Bourg, Florent Daigniere, Michael Szediwy, Michael Peterson, Markus Kilås, Erwin Tratar and Björn Kautler.</p>
 
 <h3>Contact</h3>
