+++ jsign-ant/src/main/resources/net/jsign/antlib.xml
@@ -1,4 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <antlib>
   <taskdef name="signexe" classname="net.jsign.PESignerTask" />
+  <taskdef name="signps" classname="net.jsign.PSSignerTask" />
 </antlib>
