+++ jsign/src/deb/data/usr/share/jsign/pssign.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+java -Xmx128m -Djava.net.useSystemProxies=true -cp /usr/share/jsign/jsign-@VERSION@.jar net.jsign.PSSignerCLI "$@"
