+++ jsign/src/deb/data/usr/share/man/man1/pesign.1
@@ -77,6 +77,10 @@ The name of the application. This is an optional information embedded in the sig
 The URL of the application. This is an optional information embedded in the signature.
 
 .TP
+.B -replace
+Tells if previous signatures should be replaced.
+
+.TP
 .B -h, --help
 Print the help
